feat: add Grunt recipe and export writeResultFiles
Document a custom Grunt task replacing archived grunt-webfont, extract
CLI disk writes into a shared module, and add a recipe smoke test.

Related issue #771
