+++ MIGRATION.md
@@ -48,7 +48,7 @@ webfont "fonts/*.ttf" -f woff2,woff,eot,svg -d dist --dest-create
 
 ### Comparison with `grunt-webfont`
 
-[`grunt-webfont`](https://github.com/sapegin/grunt-webfont) is the classic "SVG → webfont" **Grunt plugin** (~1.1k★, MIT). It is **archived** (read-only since 2021, last release `v1.2.0`), and its best-fidelity output relies on a native **FontForge** binary (its pure-`node` engine can't do ligatures). `webfont` is a framework-agnostic, actively maintained, pure-JS library **and** CLI: it does everything grunt-webfont *generated* — plus TTF encoding, WOFF/WOFF2 decompression, more templates, and programmatic hooks. The one thing grunt-webfont has that `webfont` doesn't is native Grunt wiring (tracked in [#771](https://github.com/itgalaxy/webfont/issues/771)); you can already drive `webfont()` from a small custom Grunt task.
+[`grunt-webfont`](https://github.com/sapegin/grunt-webfont) is the classic "SVG → webfont" **Grunt plugin** (~1.1k★, MIT). It is **archived** (read-only since 2021, last release `v1.2.0`), and its best-fidelity output relies on a native **FontForge** binary (its pure-`node` engine can't do ligatures). `webfont` is a framework-agnostic, actively maintained, pure-JS library **and** CLI: it does everything grunt-webfont *generated* — plus TTF encoding, WOFF/WOFF2 decompression, more templates, and programmatic hooks. For Grunt, use the documented [custom task recipe](../../docs/recipes/grunt-webfont/README.md) ([#771](https://github.com/itgalaxy/webfont/issues/771)).
 
 | Capability | `webfont` (this project) | `grunt-webfont` ([sapegin/grunt-webfont](https://github.com/sapegin/grunt-webfont)) |
 |------------|:------------------------:|:----------------------------------------------------------------------------------:|
@@ -67,6 +67,6 @@ webfont "fonts/*.ttf" -f woff2,woff,eot,svg -d dist --dest-create
 | Codepoint map persistence | ❌ | ✅ (`codepointsFile`) |
 | Programmatic API | ✅ (`webfont()`) | ❌ (Grunt only) |
 | SVG diagnostics | ✅ (alpha) | ❌ |
-| Grunt integration | ⏳ ([#771](https://github.com/itgalaxy/webfont/issues/771)) | ✅ (native) |
+| Grunt integration | ✅ ([recipe](../../docs/recipes/grunt-webfont/README.md)) | ✅ (native) |
 
 If you only need the generation that grunt-webfont did, `webfont` is a maintained, FontForge-free replacement. If you specifically need BEM/Bootstrap CSS presets or `data:uri` embedding today, use a [custom template](../../packages/webfont/docs/configuration.md#template) in the configuration guide.
