+++ docs/recipes/grunt-webfont/package.json
@@ -1,0 +1,13 @@
+{
+  "private": true,
+  "name": "grunt-webfont-recipe",
+  "description": "Smoke-test fixture for the webfont Grunt recipe (not published)",
+  "scripts": {
+    "grunt": "grunt"
+  },
+  "devDependencies": {
+    "grunt": "1.6.1",
+    "grunt-cli": "1.5.0",
+    "webfont": "file:../../../packages/webfont"
+  }
+}
